refactor: Extract RpcPalette into tui::rpc_palette module
