+++ twinleaf-tools/src/tui/mod.rs
@@ -1,0 +1,1 @@
+pub mod rpc_palette;
